[visualize-api-return] creation of entity class
Cause: There is no way to visualize the API return values.

Changes: A TrainingResponse class was created in data/model, alongside classes Usage and Choice, all of which are meant to be direct models of the api response to be mapped. Small changes were also made to the Message class by placing the constant ROLE in RepositoryImpl, to increase versatility.